Closes#1
Here are the changes I made:

- Reduced DOM accesses:
Instead of checking the element types (checkbox, radio, select, textarea, input) multiple times in different places, I created separate functions handleFieldLoad and handleFieldSave to handle these checks. This helps avoid duplication of logic and reduces the number of DOM accesses.

-Single Responsibility Principle:
The loadFormData function now delegates the task of updating field values to the handleFieldLoad function, which is responsible for handling the logic of setting the value based on the element type.
The saveFormData function delegates the logic of saving field values to the handleFieldSave function, which determines whether to save the checkbox/radio checked state or the regular field value.

These changes improve the code by making it cleaner, more efficient, and easier to maintain.